requirements.txt

streamlit
python-docx
langchain
faiss-cpu
sentence-transformers
llama-cpp-python
python-dotenv
requests
